+++ src/prototypes.hpp
@@ -296,8 +296,8 @@
         //Valid Theme Colors
         enum class Colors: int{ GossipTxt = 0, ShoutTxt, EmoteTxt, DeathTxt, SpellTxt, DiceAndLevel, PlayerTxt, AdminTxt, AdminMessage, OwnerTxt, CommentTxt,
                                 GoldenSoul, WhiteSoul, PlayerName, AdminName, OwnerName, TimeStamp, AdminValid, AdminInvalid, IPValid,
-                                IPInvalid, IPVanished, PartyJoin, PKChallenge, SoulIncarnated, SoulLeftWorld = 25,
-                                ColorCount = 25, Default = -1 };
+                                IPInvalid, IPQuarantined, IPVanished, PartyJoin, PKChallenge, SoulIncarnated, SoulLeftWorld = 26,
+                                ColorCount = 26, Default = -1 };
 
         //Valid columns within the PlrListWidget.
         enum class PlrCols: int{ IPPort = 0, SerNum, Age, Alias, Time,